% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_family.R
\name{get_family}
\alias{get_family}
\title{Retrieve the taxonomic families of a tibble containing genus and species
pairings}
\usage{
get_family()
}
\description{
Retrieve the taxonomic families of a tibble containing genus and species
pairings
}
